■準委任作業 　・No.2 保全活動詳細画面で機器別管理基準単位で作業記録ができるようにする
　・No.3 長期計画、保全活動 長期計画から保全活動を作成する際に機器別管理基準の行単位で保全活動を作成できるようにしたい

　・No.4 帳票ー機器台帳詳細(RP0020)　で機種仕様を出力できるようにする。

■移行課題
　・No.353 ユーザ権限が正しく設定されない

　・No.374 保全実績評価表示時に初期化エラーが発生

　・No.379 対象機器から削除した機器台帳が削除されないまま残ってしまう

　・No.387 特定工場で対象機器別管理基準を検索時にエラーが発生する

■その他
　・ExcelPort機器台帳でメーカーを空白に更新可能とする
　　※準委任作業2024年6月度 の作業項目No.19(2024_06_実施リスト.xlsx より)

■マニュアル(2024/05/24 に連携済み)
　・ファイルやりとり専用 > 納品物 > 20240524_準委任作業2024年05月度(マニュアルのみ)

+++ TMQ6WebSolution/BackEndSolution/BusinessLogic_MP0001/BusinessLogic_MP0001.cs
@@ -1667,7 +1667,7 @@ namespace BusinessLogic_MP0001
                 {
                     data.Other = "0"; // その他
                 }
-                int total = int.Parse(data.Machine.Replace(",", "")) + int.Parse(data.Electricity.Replace(",", "")) + int.Parse(data.Instrumentation.Replace(",", "")) + int.Parse(data.Other.Replace(",", ""));
+                double total = double.Parse(data.Machine.Replace(",", "")) + double.Parse(data.Electricity.Replace(",", "")) + double.Parse(data.Instrumentation.Replace(",", "")) + double.Parse(data.Other.Replace(",", ""));
                 if (total != 0)
                 {
                     return true;
